+++ components/core/LoginModalButton.tsx
@@ -9,12 +9,14 @@ import {
   Spinner,
   Flex,
   ModalHeader,
+  Stack,
 } from '@chakra-ui/react';
 import { FC } from 'react';
 import { ActionButton } from '../ActionButton';
 import { LoginComponent } from './LoginComponent';
 import { useEffectOnlyOnUpdate } from '../../hooks/useEffectOnlyOnUpdate';
-import { useLogin, useLogout } from '@useelven/core';
+import { getSigningDeviceName } from '../../utils/getSigningDeviceName';
+import { useLogin, useLogout, useLoginInfo } from '@useelven/core';
 
 interface LoginModalButtonProps {
   onClose?: () => void;
@@ -29,7 +31,8 @@ export const LoginModalButton: FC<LoginModalButtonProps> = ({
   onClose,
   onOpen,
 }) => {
-  const { isLoggedIn, isLoggingIn } = useLogin();
+  const { isLoggedIn, isLoggingIn, setLoggingInState } = useLogin();
+  const { loginMethod } = useLoginInfo();
   const { logout } = useLogout();
   const {
     isOpen: opened,
@@ -42,6 +45,12 @@ export const LoginModalButton: FC<LoginModalButtonProps> = ({
       close();
     }
   }, [isLoggedIn]);
+
+  const onCloseComplete = () => {
+    setLoggingInState('error', '');
+  };
+
+  const ledgerOrPortalName = getSigningDeviceName(loginMethod);
 
   return (
     <>
@@ -56,6 +65,7 @@ export const LoginModalButton: FC<LoginModalButtonProps> = ({
         onClose={close}
         isCentered
         scrollBehavior="inside"
+        onCloseComplete={onCloseComplete}
       >
         <CustomModalOverlay />
         <ModalContent
@@ -80,13 +90,22 @@ export const LoginModalButton: FC<LoginModalButtonProps> = ({
                 justifyContent="center"
                 position="absolute"
                 inset={0}
+                zIndex="overlay"
               >
-                <Spinner
-                  thickness="3px"
-                  speed="0.4s"
-                  color="elvenTools.color2.base"
-                  size="xl"
-                />
+                <Stack alignItems="center">
+                  {ledgerOrPortalName ? (
+                    <>
+                      <Text fontSize="lg">Confirmation required</Text>
+                      <Text fontSize="sm">Approve on {ledgerOrPortalName}</Text>
+                    </>
+                  ) : null}
+                  <Spinner
+                    thickness="3px"
+                    speed="0.4s"
+                    color="elvenTools.color2.base"
+                    size="xl"
+                  />
+                </Stack>
               </Flex>
             )}
             <LoginComponent />