+++ pages/_app.tsx
@@ -24,6 +24,9 @@ const ElvenToolsDapp = ({ Component, pageProps }: AppProps) => {
     ...(process.env.NEXT_PUBLIC_MULTIVERSX_API
       ? { apiAddress: process.env.NEXT_PUBLIC_MULTIVERSX_API }
       : {}),
+    ...(process.env.NEXT_PUBLIC_WC_PROJECT_ID
+      ? { walletConnectV2ProjectId: process.env.NEXT_PUBLIC_WC_PROJECT_ID }
+      : {}),
   });
 
   const toast = useToast();